Update sample plugins for compatibility with gmxapi 0.0.7 as defined by GROMACS 2019.
Updates to development environment:

* require gmxapi version 0.0.7
* `site` type changed from `unsigned long int` to `int`
* updates for change of gmx::Vector type alias from gmx::vec3<real> to
  gmx::RVec, which has no negation operator and requires arithmetic
  operations involving scalars to have explicitly `real` type.
* add missing virtual destructor to classes derived from
  gmx::IRestraintPotential
* IRestraintPotential::name() becomes a const member function.

Additional updates to example code:

* add trailing underscore ('_') to data member names to help clarify
  variable scope.
* remove (currently) unused references to
  `gmxapi/session/outputstream.h`
* include some potentially missing cassert headers.
* use `EXPECT_FLOAT_EQ(0., norm(vector1 - vector2))` instead of equality
  operator for testing RVec equivalence.

Updates to build and test infrastructure:

* replace install_gromacs_0_0_6.sh with install_gromacs_2019.sh
* replace pygmx_0_0_6.sh with pygmx_0_0_7.sh
* explicitly set GMX_ENABLE_CCACHE=ON
* clean up travis.yml
* don't redundantly pull repository in sample_restraint.sh
* use `EXPECT` instead of `ASSERT` test macros for expedient testing.
* remove some outdated testing.

Temporary changes point to non-official branch tips for GROMACS and for
the gmxapi Python package.

+++ src/pythonmodule/export_plugin.cpp
@@ -12,9 +12,12 @@
 
 #include "export_plugin.h"
 
+#include <cassert>
+
 #include "pybind11/pybind11.h"
 #include "pybind11/stl.h"
 
+#include "gmxapi/exceptions.h"
 #include "gmxapi/md.h"
 #include "gmxapi/md/mdmodule.h"
 #include "gmxapi/gmxapi.h"
@@ -210,8 +213,8 @@ class HarmonicRestraintBuilder
             assert(parameter_dict.contains("R0"));
             assert(parameter_dict.contains("k"));
             py::list sites = parameter_dict["sites"];
-            site1Index_ = py::cast<unsigned long>(sites[0]);
-            site2Index_ = py::cast<unsigned long>(sites[1]);
+            site1Index_ = py::cast<int>(sites[0]);
+            site2Index_ = py::cast<int>(sites[1]);
             equilibriumPosition_ = py::cast<real>(parameter_dict["R0"]);
             springConstant_ = py::cast<real>(parameter_dict["k"]);
         };
@@ -254,8 +257,8 @@ class HarmonicRestraintBuilder
         };
 
         py::object subscriber_;
-        unsigned long site1Index_;
-        unsigned long site2Index_;
+        int site1Index_;
+        int site2Index_;
         real equilibriumPosition_;
         real springConstant_;
 };
@@ -282,7 +285,7 @@ class EnsembleRestraintBuilder
             py::list sites = parameter_dict["sites"];
             for (auto&& site : sites)
             {
-                siteIndices_.emplace_back(py::cast<unsigned long>(site));
+                siteIndices_.emplace_back(py::cast<int>(site));
             }
 
             auto nbins = py::cast<size_t>(parameter_dict["nbins"]);
@@ -380,7 +383,7 @@ class EnsembleRestraintBuilder
 
         py::object subscriber_;
         py::object context_;
-        std::vector<unsigned long int> siteIndices_;
+        std::vector<int> siteIndices_;
 
         plugin::ensemble_input_param_type params_;
 
@@ -512,8 +515,8 @@ PYBIND11_MODULE(myplugin, m) {
     // Deprecated constructor directly taking restraint paramaters.
     harmonic.def(
         py::init(
-            [](unsigned long int site1,
-               unsigned long int site2,
+            [](int site1,
+               int site2,
                real R0,
                real k) {
                 return PyRestraint<plugin::HarmonicModule>::create(site1,
